+++ lib/pipetree/inspect.rb
@@ -10,18 +10,8 @@ module Pipetree::Inspect
     inspect_rows(names)
   end
 
-  # open source file to retrieve the constant name.
   def inspect_func(func)
-    return inspect_object(func) unless func.is_a?(Proc)
-    inspect_proc(func)
-  end
-
-  def inspect_object(obj)
-    obj.inspect.sub(/0x\w+/, "")
-  end
-
-  def inspect_proc(proc)
-    File.readlines(proc.source_location[0])[proc.source_location[1]-1].match(/^\s+([\w\:]+)/)[1]
+    func
   end
 
   def inspect_line(names)