+++ CHANGES.md
@@ -1,8 +1,11 @@
 # 0.0.6
 
+* All "operators" like `Flow#&` are now optional in `Operator` and considered experimental. Single entry point to adding steps is `Flow#add`.
 * Remove `Flow#>>`.
-* Add `Flow#add` which allows a low-level *strut*, a step that sits directly on the track with incoming interface `(last, input, options)` and outgoing `[new_track, input]`.
+* Add `Flow#add` which allows a low-level *tie*, a step that sits directly on the track with incoming interface `(last, input, options)` and outgoing `[new_track, input]`.
+* Make `Flow#_insert` private, as we now have `add`.
 * `And`'s direction can now be configured via `:on_true` and `:on_false`, allowing to deviate to any kind of track.
+* Remove the `Pipetree` array implementation and save it for future versions.
 
 # 0.0.5